requirements/base.txt
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-environ==0.4.5
django-jet==1.0.8
django-leaflet==0.27.1
djangorestframework==3.11.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
Pillow==7.2.0
protobuf==3.12.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyQRCode==1.2.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
XlsxWriter==1.3.6
openpyxl==3.0.5